rename studio.py to rt_studio.py
